refactor : 광고 리포지터리 구조 뷰에서 도메인으로 이동
