+++ app/src/main/java/woowacourse/movie/ui/main/movieList/MovieListPresenter.kt
@@ -1,7 +1,8 @@
 package woowacourse.movie.ui.main.movieList
 
-import woowacourse.movie.data.AdvRepository
-import woowacourse.movie.data.MovieRepository
+import com.example.domain.repository.AdvRepository
+import woowacourse.movie.model.mapper.asPresentation
+import woowacourse.movie.repository.MovieRepository
 
 class MovieListPresenter(
     private val view: MovieListContract.View,
@@ -9,6 +10,9 @@ class MovieListPresenter(
     private val advRepository: AdvRepository = AdvRepository
 ) : MovieListContract.Presenter {
     override fun getAdapter() {
-        view.setAdapter(movieRepository.allMovies(), advRepository.allAdv())
+        view.setAdapter(
+            movieRepository.allMovies(),
+            advRepository.allAdv().map { it.asPresentation() }
+        )
     }
 }
