+++ app/src/main/res/layout/activity_adv_detail.xml
@@ -29,8 +29,8 @@
                     android:layout_width="match_parent"
                     android:layout_height="250dp"
                     android:adjustViewBounds="true"
-                    android:imageResource="@{adv.imgId}"
                     android:scaleType="centerInside"
+                    android:src="@{adv.imageUri}"
                     app:layout_constraintBottom_toTopOf="@id/adv_description"
                     app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintStart_toStartOf="parent"
