+++ app/src/main/res/layout/item_adv_layout.xml
@@ -3,8 +3,8 @@
     xmlns:android="http://schemas.android.com/apk/res/android">
     <data>
         <variable
-            name="imageId"
-            type="int" />
+            name="adv"
+            type="woowacourse.movie.model.AdvState" />
     </data>
 <LinearLayout android:layout_width="match_parent"
     android:layout_height="wrap_content"
@@ -14,7 +14,7 @@
         android:id="@+id/adv_img"
         android:layout_width="match_parent"
         android:layout_height="96dp"
-        android:imageResource="@{imageId}"
+        android:src="@{adv.imageUri}"
         android:adjustViewBounds="true"
         android:scaleType="centerCrop"
         tools:ignore="ContentDescription" />
