+++ domain/src/main/java/com/example/domain/model/Adv.kt
@@ -1,0 +1,6 @@
+package com.example.domain.model
+
+data class Adv(
+    val imageUrl: String,
+    val advDescription: String
+)
